tools/assets/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
noise>=1.2.2
